Simulation 6.1.
Added some minor tweaks to the simulation.

-lower initial stock
-higher restock amount

Considering Adding another feature:
Supply Chain Issue. Where you either get lower amount of restocking or no restocking. Problems currently are that it would influence the restocking amount a lot and restocking as  a whole would have to be tweaked in order to balance those out but it could improve complexity.